Implementacion api Rest albaranes y proveedores

+++ src/main/java/com/example/server/UserRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import pojos.Usuarios;
 
-public interface UserRepository extends JpaRepository <Usuarios , Long> {
+public interface UserRepository extends JpaRepository <Usuarios ,Integer> {
 	Usuarios findByEmail(String email);
 
 	Usuarios findByNombre(String nombre);
